+++ modules/System/views/api/rest-api-viewer.php
@@ -5,7 +5,7 @@
 </head>
 <body>
     <rapi-doc
-        spec-url="<?=$openApiUrl?>"
+        spec-url="<?=$this->escape($openApiUrl)?>"
         show-header="false"
         show-info="false"
         render-style="read"
@@ -14,12 +14,12 @@
         <?php if($apiKey): ?>
         api-key-name = "api-key"
         api-key-location = "header"
-        api-key-value = "<?=$apiKey?>"
+        api-key-value = "<?=$this->escape($apiKey)?>"
         <?php endif ?>
 
-        bg-color="<?=($bgColor ? $bgColor : '#10131a')?>"
-        text-color="<?=($textColor ? $textColor : '#fafafa')?>"
-        primary-color="<?=($primaryColor ? $primaryColor : '#0e8fff')?>"
+        bg-color="<?=$this->escape($bgColor ? $bgColor : '#10131a')?>"
+        text-color="<?= $this->escape($textColor ? $textColor : '#fafafa')?>"
+        primary-color="<?= $this->escape($primaryColor ? $primaryColor : '#0e8fff')?>"
     ></rapi-doc>
 
     <script type="module" src="<?=$this->base('system:assets/vendor/rapidoc.js')?>"></script>