+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## WIP
 
 - Fix mime type detection on asset upload
+- Fix possible Cross-site Scripting (XSS) in Rest/GraphQL viewer
 
 ## 2.6.3 (2023-08-15)